… src/mscorlib/src/System/Reflection/CustomAttribute.cs
… src/mscorlib/src/System/RtType.cs