@benaadams 2016-08-07
Shame you can't do something like
```csharp
public static class EnumNonBoxExtensions
{
    static unsafe string ToString<TEnum>(this TEnum e) where TEnum : Enum
    {
        fixed (void* pValue = &JitHelpers.GetPinningHelper(e).m_data)
        {
            // ...
        }
    }
}
```