@jkotas 2016-08-08
LGTM modulo comments. Thank you for the perf improvement!