@jkotas 2016-08-08
Could you please run the corefx tests on this? I see number of failures, like:

            System.Tests.EnumTests.Format(enumType: typeof(System.Tests.EnumTests+SimpleEnum), value: 1, format: "X", expected: "00000001") [FAIL]
               System.InvalidCastException : Unable to cast object of type 'System.Int32' to type 'System.Enum'.