@benaadams 2016-08-09
I'm not entirely sure how to run the corefx tests against coreclr, the instructions to override the runtime don't appear to be correct anymore https://github.com/dotnet/coreclr/blob/master/Documentation/building/testing-with-corefx.md

I'm copying the tests into a separate standalone exe and converting them to functions, I have most of them converted; though haven't finished them all yet to be sure. (found and fixed a bug in the Enum)