@jkotas 2016-08-09
The msbuild arguments have to be passed after `--` after recent changes by @maririos 

`build.cmd -- /p:BUILDTOOLS_OVERRIDE_RUNTIME:...`