@jkotas 2016-08-09
I think there is a bug with negative values:

```
using System;

enum MyEnum {
    MyValue = -1
}

class My {
    static void Main() {
        MyEnum x = MyEnum.MyValue;
        Console.WriteLine(x.ToString());
    }
}
```

Actual: -1
Expected: MyValue

If there is no test coverage for this case in corefx, could you please add it?