@jackfree 2016-08-09
@benaadams thank you tremendously for making these changes!