@benaadams 2016-08-10
@jasonwilliams200OK reset of the time is elsewhere

![enum-remain](https://aoa.blob.core.windows.net/aspnet/enum-remain.png)