@benaadams 2016-08-10
Or perhaps better
![](https://aoa.blob.core.windows.net/aspnet/enum-remain-2.png)