@benaadams 2016-08-10
e.g. 76.6% of the time is working out if its a flag Enum

![flag-check](https://aoa.blob.core.windows.net/aspnet/enum-remain-3.png)