@benaadams 2016-08-10
Might be able to mitigate that....