@benaadams 2016-08-10
Added whether its Flags to the cache