@benaadams 2016-08-10
@jkotas @jackfree @jasonwilliams200OK Woop! Updated the numbers, its now 7x faster than the current...